+++ packages/polaris-viz/src/components/LineChartRelational/components/MissingDataArea/tests/MissingDataArea.test.tsx
@@ -10,35 +10,44 @@ const MOCK_PROPS: Props = {
       name: 'Apr 1 – Apr 14, 2020',
       data: [
         {value: 333, key: '2020-04-01T12:00:00'},
-        {value: 797, key: '2020-04-02T12:00:00'},
+        {value: 333, key: '2020-04-02T12:00:00'},
         {value: 234, key: '2020-04-03T12:00:00'},
-        {value: 534, key: '2020-04-04T12:00:00'},
+        {value: 333, key: '2020-04-04T12:00:00'},
+        {value: 333, key: '2020-04-05T12:00:00'},
       ],
     },
     {
       name: '75th Percentile',
       data: [
         {value: 333, key: '2020-04-01T12:00:00'},
-        {value: 797, key: '2020-04-02T12:00:00'},
+        {value: 333, key: '2020-04-02T12:00:00'},
+        {value: 333, key: '2020-04-03T12:00:00'},
+        {value: null, key: '2020-04-04T12:00:00'},
+        {value: null, key: '2020-04-04T12:00:00'},
       ],
     },
     {
       name: 'Similar stores median',
       data: [
         {value: 333, key: '2020-04-01T12:00:00'},
-        {value: 797, key: '2020-04-02T12:00:00'},
+        {value: 333, key: '2020-04-02T12:00:00'},
+        {value: 333, key: '2020-04-03T12:00:00'},
+        {value: null, key: '2020-04-04T12:00:00'},
+        {value: null, key: '2020-04-04T12:00:00'},
       ],
     },
     {
       name: '25th percentile',
       data: [
         {value: 333, key: '2020-04-01T12:00:00'},
-        {value: 797, key: '2020-04-02T12:00:00'},
+        {value: 333, key: '2020-04-02T12:00:00'},
+        {value: 333, key: '2020-04-03T12:00:00'},
+        {value: null, key: '2020-04-04T12:00:00'},
+        {value: null, key: '2020-04-04T12:00:00'},
       ],
     },
   ],
   drawableHeight: 200,
-  drawableWidth: 600,
   xScale: scaleLinear(),
 };
 
@@ -50,10 +59,13 @@ describe('<MissingDataArea />', () => {
       </svg>,
     );
 
-    expect(chart).toContainReactComponent('rect');
+    expect(chart).toContainReactComponent('rect', {
+      x: 2,
+      width: 2,
+    });
   });
 
-  it('renders nothing when data series have same length', () => {
+  it('renders nothing when data has no nulls', () => {
     const chart = mount(
       <svg>
         <MissingDataArea
@@ -63,20 +75,29 @@ describe('<MissingDataArea />', () => {
               name: 'Apr 1 – Apr 14, 2020',
               data: [
                 {value: 333, key: '2020-04-01T12:00:00'},
-                {value: 797, key: '2020-04-02T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
               ],
             },
             {
               name: '75th Percentile',
-              data: [],
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
+              ],
             },
             {
               name: 'Similar stores median',
-              data: [],
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
+              ],
             },
             {
               name: '25th percentile',
-              data: [],
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
+              ],
             },
           ]}
         />
@@ -84,5 +105,129 @@ describe('<MissingDataArea />', () => {
     );
 
     expect(chart).not.toContainReactComponent('rect');
+  });
+
+  it('renders area at the start of the chart', () => {
+    const chart = mount(
+      <svg>
+        <MissingDataArea
+          {...MOCK_PROPS}
+          data={[
+            {
+              name: 'Apr 1 – Apr 14, 2020',
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
+                {value: 333, key: '2020-04-03T12:00:00'},
+                {value: 333, key: '2020-04-04T12:00:00'},
+                {value: 333, key: '2020-04-05T12:00:00'},
+              ],
+            },
+            {
+              name: '75th Percentile',
+              data: [
+                {value: null, key: '2020-04-01T12:00:00'},
+                {value: null, key: '2020-04-02T12:00:00'},
+                {value: null, key: '2020-04-03T12:00:00'},
+                {value: 333, key: '2020-04-04T12:00:00'},
+                {value: 333, key: '2020-04-05T12:00:00'},
+              ],
+            },
+          ]}
+        />
+      </svg>,
+    );
+
+    expect(chart).toContainReactComponent('rect', {
+      x: 0,
+      width: 3,
+    });
+  });
+
+  it('renders areas in the middle of the chart', () => {
+    const chart = mount(
+      <svg>
+        <MissingDataArea
+          {...MOCK_PROPS}
+          data={[
+            {
+              name: 'Apr 1 – Apr 14, 2020',
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
+                {value: 333, key: '2020-04-03T12:00:00'},
+                {value: 333, key: '2020-04-04T12:00:00'},
+                {value: 333, key: '2020-04-05T12:00:00'},
+                {value: 333, key: '2020-04-06T12:00:00'},
+                {value: 333, key: '2020-04-07T12:00:00'},
+                {value: 333, key: '2020-04-08T12:00:00'},
+                {value: 333, key: '2020-04-09T12:00:00'},
+              ],
+            },
+            {
+              name: '75th Percentile',
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
+                {value: 333, key: '2020-04-03T12:00:00'},
+                {value: null, key: '2020-04-04T12:00:00'},
+                {value: null, key: '2020-04-05T12:00:00'},
+                {value: null, key: '2020-04-06T12:00:00'},
+                {value: 333, key: '2020-04-07T12:00:00'},
+                {value: 333, key: '2020-04-08T12:00:00'},
+                {value: 333, key: '2020-04-09T12:00:00'},
+              ],
+            },
+          ]}
+        />
+      </svg>,
+    );
+
+    expect(chart).toContainReactComponent('rect', {
+      x: 2,
+      width: 4,
+    });
+  });
+
+  it('renders areas at random areas', () => {
+    const chart = mount(
+      <svg>
+        <MissingDataArea
+          {...MOCK_PROPS}
+          data={[
+            {
+              name: 'Apr 1 – Apr 14, 2020',
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: 333, key: '2020-04-02T12:00:00'},
+                {value: 333, key: '2020-04-03T12:00:00'},
+                {value: 333, key: '2020-04-04T12:00:00'},
+                {value: 333, key: '2020-04-05T12:00:00'},
+                {value: 333, key: '2020-04-06T12:00:00'},
+                {value: 333, key: '2020-04-07T12:00:00'},
+                {value: 333, key: '2020-04-08T12:00:00'},
+                {value: 333, key: '2020-04-09T12:00:00'},
+              ],
+            },
+            {
+              name: '75th Percentile',
+              data: [
+                {value: 333, key: '2020-04-01T12:00:00'},
+                {value: null, key: '2020-04-02T12:00:00'},
+                {value: null, key: '2020-04-03T12:00:00'},
+                {value: 333, key: '2020-04-04T12:00:00'},
+                {value: 333, key: '2020-04-05T12:00:00'},
+                {value: 333, key: '2020-04-06T12:00:00'},
+                {value: null, key: '2020-04-07T12:00:00'},
+                {value: null, key: '2020-04-08T12:00:00'},
+                {value: 333, key: '2020-04-09T12:00:00'},
+              ],
+            },
+          ]}
+        />
+      </svg>,
+    );
+
+    expect(chart).toContainReactComponentTimes('rect', 2);
   });
 });