+++ packages/polaris-viz/src/components/LineChartRelational/stories/chromatic/data/missing-end-data.js
@@ -1,67 +1,4 @@
-import type {Story} from '@storybook/react';
-import type {RenderTooltipContentData} from 'types';
-import type {DataSeries} from '@shopify/polaris-viz-core';
-import type {LineChartProps} from 'components/LineChart/LineChart';
-
-import {LineChartRelational} from '../LineChartRelational';
-import {
-  formatLinearXAxisLabel,
-  formatLinearYAxisLabel,
-} from '../../../storybook/utilities';
-import {renderLinearTooltipContent} from '../../../utilities';
-
-export const TOOLTIP_CONTENT = {
-  empty: undefined,
-  Custom: ({data}: RenderTooltipContentData) => {
-    return (
-      <div
-        style={{
-          background: 'black',
-          color: 'white',
-          padding: '10px',
-          borderRadius: '10px',
-          display: 'flex',
-          flexDirection: 'column',
-          fontSize: 12,
-        }}
-      >
-        {data[0].data.map(({key, value}) => (
-          // eslint-disable-next-line @shopify/jsx-no-hardcoded-content
-          <div key={key}>{`${key}: ${formatLinearYAxisLabel(
-            Number(value!),
-          )}`}</div>
-        ))}
-      </div>
-    );
-  },
-};
-
-export const Template: Story<LineChartProps> = (args: LineChartProps) => {
-  return <LineChartRelational {...args} />;
-};
-
-export const DEFAULT_PROPS: Partial<LineChartProps> = {
-  xAxisOptions: {
-    labelFormatter: formatLinearXAxisLabel,
-  },
-  yAxisOptions: {labelFormatter: formatLinearYAxisLabel},
-  tooltipOptions: {
-    titleFormatter: (value) => new Date(value!).toLocaleDateString(),
-    valueFormatter: formatLinearYAxisLabel,
-    renderTooltipContent: (tooltipData) => {
-      return renderLinearTooltipContent(tooltipData, {
-        title: tooltipData.title,
-        groups: [
-          {title: 'Your store', indexes: [0]},
-          {title: 'Similar stores', indexes: [1, 2, 3]},
-        ],
-      });
-    },
-  },
-  showLegend: false,
-};
-
-export const DEFAULT_DATA: DataSeries[] = [
+export const MISSING_END_DATA = [
   {
     name: 'Average',
     data: [
@@ -98,10 +35,10 @@ export const DEFAULT_DATA: DataSeries[] = [
       {value: 162, key: '2020-03-08T12:00:00'},
       {value: 540, key: '2020-03-09T12:00:00'},
       {value: 193, key: '2020-03-10T12:00:00'},
-      {value: 860, key: '2020-03-11T12:00:00'},
-      {value: 941, key: '2020-03-12T12:00:00'},
-      {value: 773, key: '2020-03-13T12:00:00'},
-      {value: 171, key: '2020-03-14T12:00:00'},
+      {value: null, key: '2020-03-11T12:00:00'},
+      {value: null, key: '2020-03-12T12:00:00'},
+      {value: null, key: '2020-03-13T12:00:00'},
+      {value: null, key: '2020-03-14T12:00:00'},
     ],
     color: 'rgba(103, 197, 228, 1)',
     metadata: {
@@ -127,10 +64,10 @@ export const DEFAULT_DATA: DataSeries[] = [
       {value: 12, key: '2020-03-08T12:00:00'},
       {value: 390, key: '2020-03-09T12:00:00'},
       {value: 43, key: '2020-03-10T12:00:00'},
-      {value: 710, key: '2020-03-11T12:00:00'},
-      {value: 791, key: '2020-03-12T12:00:00'},
-      {value: 623, key: '2020-03-13T12:00:00'},
-      {value: 21, key: '2020-03-14T12:00:00'},
+      {value: null, key: '2020-03-11T12:00:00'},
+      {value: null, key: '2020-03-12T12:00:00'},
+      {value: null, key: '2020-03-13T12:00:00'},
+      {value: null, key: '2020-03-14T12:00:00'},
     ],
     color: 'rgba(40, 106, 123, 1)',
     metadata: {
@@ -157,10 +94,10 @@ export const DEFAULT_DATA: DataSeries[] = [
       {value: 0, key: '2020-03-08T12:00:00'},
       {value: 240, key: '2020-03-09T12:00:00'},
       {value: 0, key: '2020-03-10T12:00:00'},
-      {value: 540, key: '2020-03-11T12:00:00'},
-      {value: 641, key: '2020-03-12T12:00:00'},
-      {value: 473, key: '2020-03-13T12:00:00'},
-      {value: 0, key: '2020-03-14T12:00:00'},
+      {value: null, key: '2020-03-11T12:00:00'},
+      {value: null, key: '2020-03-12T12:00:00'},
+      {value: null, key: '2020-03-13T12:00:00'},
+      {value: null, key: '2020-03-14T12:00:00'},
     ],
     color: 'rgba(103, 197, 228, 1)',
     styleOverride: {