+++ packages/polaris-viz/src/components/LineChartRelational/stories/chromatic/data/missing-middle-data.js
@@ -1,67 +1,4 @@
-import type {Story} from '@storybook/react';
-import type {RenderTooltipContentData} from 'types';
-import type {DataSeries} from '@shopify/polaris-viz-core';
-import type {LineChartProps} from 'components/LineChart/LineChart';
-
-import {LineChartRelational} from '../LineChartRelational';
-import {
-  formatLinearXAxisLabel,
-  formatLinearYAxisLabel,
-} from '../../../storybook/utilities';
-import {renderLinearTooltipContent} from '../../../utilities';
-
-export const TOOLTIP_CONTENT = {
-  empty: undefined,
-  Custom: ({data}: RenderTooltipContentData) => {
-    return (
-      <div
-        style={{
-          background: 'black',
-          color: 'white',
-          padding: '10px',
-          borderRadius: '10px',
-          display: 'flex',
-          flexDirection: 'column',
-          fontSize: 12,
-        }}
-      >
-        {data[0].data.map(({key, value}) => (
-          // eslint-disable-next-line @shopify/jsx-no-hardcoded-content
-          <div key={key}>{`${key}: ${formatLinearYAxisLabel(
-            Number(value!),
-          )}`}</div>
-        ))}
-      </div>
-    );
-  },
-};
-
-export const Template: Story<LineChartProps> = (args: LineChartProps) => {
-  return <LineChartRelational {...args} />;
-};
-
-export const DEFAULT_PROPS: Partial<LineChartProps> = {
-  xAxisOptions: {
-    labelFormatter: formatLinearXAxisLabel,
-  },
-  yAxisOptions: {labelFormatter: formatLinearYAxisLabel},
-  tooltipOptions: {
-    titleFormatter: (value) => new Date(value!).toLocaleDateString(),
-    valueFormatter: formatLinearYAxisLabel,
-    renderTooltipContent: (tooltipData) => {
-      return renderLinearTooltipContent(tooltipData, {
-        title: tooltipData.title,
-        groups: [
-          {title: 'Your store', indexes: [0]},
-          {title: 'Similar stores', indexes: [1, 2, 3]},
-        ],
-      });
-    },
-  },
-  showLegend: false,
-};
-
-export const DEFAULT_DATA: DataSeries[] = [
+export const MISSING_MIDDLE_DATA = [
   {
     name: 'Average',
     data: [
@@ -93,9 +30,9 @@ export const DEFAULT_DATA: DataSeries[] = [
       {value: 340, key: '2020-03-03T12:00:00'},
       {value: 240, key: '2020-03-04T12:00:00'},
       {value: 387, key: '2020-03-05T12:00:00'},
-      {value: 760, key: '2020-03-07T12:00:00'},
-      {value: 122, key: '2020-03-06T12:00:00'},
-      {value: 162, key: '2020-03-08T12:00:00'},
+      {value: null, key: '2020-03-07T12:00:00'},
+      {value: null, key: '2020-03-06T12:00:00'},
+      {value: null, key: '2020-03-08T12:00:00'},
       {value: 540, key: '2020-03-09T12:00:00'},
       {value: 193, key: '2020-03-10T12:00:00'},
       {value: 860, key: '2020-03-11T12:00:00'},
@@ -122,9 +59,9 @@ export const DEFAULT_DATA: DataSeries[] = [
       {value: 190, key: '2020-03-03T12:00:00'},
       {value: 90, key: '2020-03-04T12:00:00'},
       {value: 237, key: '2020-03-05T12:00:00'},
-      {value: 580, key: '2020-03-07T12:00:00'},
-      {value: 172, key: '2020-03-06T12:00:00'},
-      {value: 12, key: '2020-03-08T12:00:00'},
+      {value: null, key: '2020-03-07T12:00:00'},
+      {value: null, key: '2020-03-06T12:00:00'},
+      {value: null, key: '2020-03-08T12:00:00'},
       {value: 390, key: '2020-03-09T12:00:00'},
       {value: 43, key: '2020-03-10T12:00:00'},
       {value: 710, key: '2020-03-11T12:00:00'},
@@ -152,9 +89,9 @@ export const DEFAULT_DATA: DataSeries[] = [
       {value: 40, key: '2020-03-03T12:00:00'},
       {value: 0, key: '2020-03-04T12:00:00'},
       {value: 87, key: '2020-03-05T12:00:00'},
-      {value: 430, key: '2020-03-07T12:00:00'},
-      {value: 22, key: '2020-03-06T12:00:00'},
-      {value: 0, key: '2020-03-08T12:00:00'},
+      {value: null, key: '2020-03-07T12:00:00'},
+      {value: null, key: '2020-03-06T12:00:00'},
+      {value: null, key: '2020-03-08T12:00:00'},
       {value: 240, key: '2020-03-09T12:00:00'},
       {value: 0, key: '2020-03-10T12:00:00'},
       {value: 540, key: '2020-03-11T12:00:00'},